Test with Java 25 and Java 21
Java 25 released September 16, 2025.  Jenkins core has supported Java 25 since Jenkins 2.534, Oct 28, 2025.

Compile and test on ci.jenkins.io with Java 25 and Java 21.

Intentionally continues to generate Java 17 byte code as configured by the plugin parent pom.

Does not compile or test with Java 17 any longer because we have found no issues in the past that were specific to the Java 17 compiler.  The plan is to drop support for Java 17 in the not too distant future so that the Jenkins project is only supporting two major Java versions at a time, Java 21 and Java 25.

Testing done:

* Confirmed that automated tests pass with Java 25